Move most classes and objects that are not Object out of object.h/cpp
